requirements.txt
djangorestframework==3.8.2
markdown==3.0.1
django-filter==2.0.0
django-rest-swagger==2.2.0
psycopg2>=2.5.4
factory-boy>=2.11.1
